requirements.txt
Pillow
numpy
tqdm
loguru
